chore(shell-security): rename security-advisor server code to shell-security
