+++ apps/web/src/lib/shell-security/posthog-tracking.ts
@@ -1,5 +1,5 @@
 /**
- * PostHog tracking for security-advisor feature.
+ * PostHog tracking for shell security feature.
  *
  * Tracks scan completions for product analytics, conversion attribution,
  * and usage reporting. Events follow the wide-event pattern used by security-agent.
@@ -11,13 +11,13 @@ import { captureException } from '@sentry/nextjs';
 
 const posthogClient = PostHogClient();
 
-type BaseSecurityAdvisorEvent = {
+type BaseShellSecurityEvent = {
   distinctId: string;
   userId: string;
   organizationId?: string;
 };
 
-type SecurityAdvisorScanCompletedEvent = BaseSecurityAdvisorEvent & {
+type ShellSecurityScanCompletedEvent = BaseShellSecurityEvent & {
   sourcePlatform: string;
   sourceMethod: string;
   pluginVersion?: string;
@@ -33,12 +33,10 @@ type SecurityAdvisorScanCompletedEvent = BaseSecurityAdvisorEvent & {
 };
 
 /**
- * Track a completed security advisor scan.
+ * Track a completed shell security scan.
  * Fired after the report is generated and the scan is recorded in the DB.
  */
-export function trackSecurityAdvisorScanCompleted(
-  properties: SecurityAdvisorScanCompletedEvent
-): void {
+export function trackShellSecurityScanCompleted(properties: ShellSecurityScanCompletedEvent): void {
   try {
     posthogClient.capture({
       distinctId: properties.distinctId,
